party-backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-dotenv
requests
sqlalchemy
aiofiles
